backend/requirements.txt
fastapi[standard]>=0.112.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymongo>=4.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
fastapi-mail>=1.4.1
PyJWT>=2.10.1
webauthn>=2.7.0 
